+++ TP4/Models/DistribucionUniforme.cs
@@ -8,17 +8,17 @@ namespace TP4.Models
 {
     internal class DistribucionUniforme
     {
-        private decimal a;
-        private decimal b;
+        private double a;
+        private double b;
 
 
-        public DistribucionUniforme(decimal _a, decimal _b)
+        public DistribucionUniforme(double _a, double _b)
         {
             this.a = _a;
             this.b = _b;
         }
 
-        public decimal generar_x_uniforme(decimal random)
+        public double generar_x_uniforme(double random)
         {
             return this.a + random * (this.b - this.a);
         }
